Created Variables and unloaded everything into Variables
