+++ src/dynamics/variables.py
@@ -1,0 +1,56 @@
+from sympy import *
+import numpy as np
+import pandas as pd
+from typing import Callable
+
+class Variables:
+    def __init__(self):
+        self.params : list = None
+        self.f_subs_params : Matrix = None
+        self.f_subs_full : Matrix = None
+        self.dt : float = None
+        self.x0 : np.ndarray = None
+        self.t0 : float = 0.0
+        self.t_sym : Symbol = None
+
+        ## Uninitialized parameters ##
+        
+        # Rocket parameters
+        self.I_0 : float = None # Initial moment of inertia in kg·m²
+        self.I_f : float = None # Final moment of inertia in kg·m²
+        self.I_3 : float = None # Rotational moment of inertia about z-axis in kg·m²
+        self.x_CG_0 : float = None # Initial center of gravity location in meters
+        self.x_CG_f : float = None # Final center of gravity location in meters
+
+
+        self.m_0 : float = None # Initial rocket mass in kg
+        self.m_f : float = None # Final rocket mass in kg
+        self.m_p : float = None # Propellant mass in kg
+        self.d : float = None # Rocket body diameter in meters
+        self.L_ne : float = None # Length from nose to nozzle in meters
+        self.C_d : float = None # Drag coefficient
+        self.Cnalpha_rocket : float = None # Rocket normal force coefficient derivative
+        self.t_motor_burnout : float = None # Time to motor burnout in seconds
+        self.t_launch_rail_clearance : float = None # Time to launch rail clearance in seconds
+        self.t_estimated_apogee : float = None # Time to apogee in esconds
+        self.CP_func : Callable[[Expr], Expr] = None # Center of pressure location as a function of angle of attack in degrees
+        
+        # Fin parameters
+        self.N : float = None # Number of fins
+        self.Cr : float = None # Root chord in meters
+        self.Ct : float = None # Tip chord in meters
+        self.s : float = None # Span in meters
+        self.delta : float = None # Fin cant angle in degrees
+        self.Cnalpha_fin : float = None # Normal FORCE coefficient normalized by angle of attack for 1 fin
+
+        
+        # Rocket name (used for saving simulation results from Simluation() object to designated path)
+        
+        
+        # State space linearization matrices
+        self.A_sym : Matrix = None # Symbolic state matrix
+        self.A : np.ndarray = None # State matrix
+        
+        ## Helpers [Refer to super for other]##
+        self._f_numeric = None  # Cached lambdified EOM
+        self._A_numeric = None  # Cached lambdified Jacobian of f wrt state
